feat: 使用MinGW64编译通过，jmp_buf使用long long的8字节

+++ md.h
@@ -71,7 +71,7 @@ typedef struct _st_jmp_buf {
      * Linux    __loongarch64           long[12]
      * Cygwin64 __amd64__/__x86_64__    long[8]
      */
-    long __jmpbuf[22];
+    long long __jmpbuf[22];
 } _st_jmp_buf_t[1];
 
 extern int _st_md_cxt_save(_st_jmp_buf_t env);
@@ -180,7 +180,7 @@ extern void _st_md_cxt_restore(_st_jmp_buf_t env, int val);
     #define MD_USE_BUILTIN_SETJMP
 
     #if defined(__amd64__) || defined(__x86_64__)
-        #define MD_GET_SP(_t) *((long *)&((_t)->context[0].__jmpbuf[6]))
+        #define MD_GET_SP(_t) *((long long *)&((_t)->context[0].__jmpbuf[6]))
     #else
         #error Unknown CPU architecture
     #endif
@@ -189,7 +189,7 @@ extern void _st_md_cxt_restore(_st_jmp_buf_t env, int val);
         ST_BEGIN_MACRO                             \
         if (MD_SETJMP((_thread)->context))         \
             _main();                                 \
-        MD_GET_SP(_thread) = (long) (_sp);         \
+        MD_GET_SP(_thread) = (long long) (_sp);         \
         ST_END_MACRO
 
     #define MD_GET_UTIME()            \
